requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==17.4.0
autocrop==1.0.2
Automat==0.6.0
autopep8==1.4.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.0
Django==3.2.18
django-cors-headers==3.1.0
django-cors-middleware==1.4.0
django-filter==2.2.0
djangorestframework==3.14.0
dlib==19.19
dramatiq[redis, watch]
gunicorn==20.0.4
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
ImageHash==4.1.0
imageio==2.5.0
incremental==16.10.1
isort==4.3.21
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
networkx==2.3
numpy==1.17.2
Pillow==6.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycrypto==2.6.1
pyinotify==0.9.6
pylint==2.3.1
pyOpenSSL==17.5.0
pyparsing==2.4.2
pyserial==3.4
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.3
pytz==2019.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.15.0
scipy==1.3.1
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
sqlparse==0.3.0
Twisted==17.9.0
typed-ast==1.4.0
urllib3==1.22
uWSGI==2.0.18
virtualenv==15.1.0
wrapt==1.11.2
